+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Security Fixes
 
 1. [#6186](https://github.com/influxdata/chronograf/pull/6186): Hardened CSRF protections on Data Explorer and unsafe query endpoints.
+1. [#6191](https://github.com/influxdata/chronograf/pull/6191): Hardened Reader-only access and read-only query enforcement.
 
 ## v1.11.0 [2026-02-19]